+++ man/examples/example-process_gmPC.R
@@ -1,4 +1,6 @@
-# Load coordinates of aligned landmarks, PC scores, and eigenvalues; vertebra index as names of array elements of aligned landmarks
+# Load coordinates of aligned landmarks, PC scores, and
+# eigenvalues; vertebra index as names of array elements
+# of aligned landmarks
 data("seal_gmm")
 
 # Process PC scores and morphological data:
@@ -6,7 +8,8 @@ seal_pca <- process_gmPC(data = seal_gmm$coord_gpa,
                          pcscores = seal_gmm$scores,
                          eigenvals = seal_gmm$eigenvals)
 
-# Process multiple datasets; specimen names in 'specimens' vector, vertebrae index in 'vertebrae' vector
+# Process multiple datasets; specimen names in 'specimens'
+# vector, vertebrae index in 'vertebrae' vector
 data("wolves_gmm")
 
 # Process PC scores and geometric morphometric dataset: